+++ src/events/events.lua
@@ -7,6 +7,7 @@ local E = Addon:GetModule("Events")
 
 local events = {
   "LootReceived",
+  "MoneyReceived",
   "TexturedLootMessage"
 }
 
@@ -19,5 +20,8 @@ end
 -- ============================================================================
 
 E.Wow = {
-  ChatMessageLoot = "CHAT_MSG_LOOT"
+  ChatMessageLoot = "CHAT_MSG_LOOT",
+  ChatMessageMoney = "CHAT_MSG_MONEY",
+  PlayerLogin = "PLAYER_LOGIN",
+  PlayerLogout = "PLAYER_LOGOUT"
 }