+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.1.0] - 2023-12-25
+
+### Added
+
+- Notifications for money when looted
+- Command: `/lootifications` to display a list of commands
+- Command: `/lootifications money` to toggle money notifications
+- Command: `/lootifications test` to test notifications
+
+### Fixed
+
+- Notification bug for loot received by other players
+
 ## [1.0.0] - 2023-12-24
 
 - Initial release